requirement.txt
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.2.0
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
cloud-init==21.4
colorama==0.4.3
colorclass==2.2.2
coloredlogs==15.0.1
colorlog==6.6.0
command-not-found==0.3
configobj==5.0.6
confluent-kafka==1.8.2
constantly==15.1.0
croniter==1.3.5
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
ec2-hibinit-agent==1.0.0
entrypoints==0.3
faust==1.10.4
frozenlist==1.3.0
gensim==4.2.0
getmac==0.8.3
h2==2.6.2
hibagent==1.0.1
hpack==3.0.0
httplib2==0.14.0
humanfriendly==10.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==19.0.0
idna==3.3
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
kafka-python==1.4.7
keyring==18.0.1
ksql==0.10.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
mode==4.3.2
more-itertools==4.2.0
multidict==6.0.2
mypy-extensions==0.4.3
netifaces==0.11.0
networkx==2.8
numpy==1.22.3
oauthlib==3.1.0
opentracing==1.3.0
pandas==1.4.2
pexpect==4.6.0
psutil==5.9.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
pythonping==1.1.1
pytz==2022.1
PyYAML==5.3.1
requests==2.27.1
requests-unixsocket==0.2.0
robinhood-aiokafka==1.1.6
scikit-learn==1.0.2
scipy==1.8.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
smart-open==6.0.0
sos==4.1
ssh-import-id==5.10
systemd-python==234
tcp-latency==0.0.11
terminaltables==3.1.10
threadpoolctl==3.1.0
tqdm==4.64.0
Twisted==18.9.0
ubuntu-advantage-tools==27.4
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
venusian==1.2.0
wadllib==1.3.3
yarl==1.7.2
zipp==1.0.0
zope.interface==4.7.1
